Updated fellows and companies table
Compared attributes on fellows and companies with the application team
and made sure we're all on the same page.

We should be now, all the fields in the application are now represented
in the database (as text or string).

Also updated the PUT calls for /companies/application/:user_id and
/fellows/application/:user_id to update the correct table columns.