Got tags working for fellows and companies - sorta
Implemented tags via PUT on /fellows/profile/:user_id and
/companies/profile/:user_id

There is one thing you have to remember, and that is you MUST call a GET
on the profile after calling the PUT. You CANNOT trust the data returned
from the PUT call. Because of some sequelize magic, we couldn't get the
tables to update before sending a response, but by the time you send
another api call for a GET the data should all be in the tables just
fine.

By the way, this only applies for when you are updating tags - if you
are updating anything else the PUT data is fine and you don't have to
call a subsequent GET.